[NFC] Factor `Result` and `MaybeResult` into a utility header
Allow them to be used for more than just the new text parser.
