+++ src/wat-parser.h
@@ -19,58 +19,10 @@
 
 #include <string_view>
 
+#include "support/result.h"
 #include "wasm.h"
 
 namespace wasm::WATParser {
-
-struct Ok {};
-
-struct None {};
-
-struct Err {
-  std::string msg;
-};
-
-template<typename T = Ok> struct Result {
-  std::variant<T, Err> val;
-
-  Result(Result<T>& other) = default;
-  Result(Result<T>&& other) = default;
-  Result(const Err& e) : val(std::in_place_type<Err>, e) {}
-  Result(Err&& e) : val(std::in_place_type<Err>, std::move(e)) {}
-  template<typename U = T>
-  Result(U&& u) : val(std::in_place_type<T>, std::forward<U>(u)) {}
-
-  Err* getErr() { return std::get_if<Err>(&val); }
-  T& operator*() { return *std::get_if<T>(&val); }
-  T* operator->() { return std::get_if<T>(&val); }
-};
-
-template<typename T = Ok> struct MaybeResult {
-  std::variant<T, None, Err> val;
-
-  MaybeResult() : val(None{}) {}
-  MaybeResult(MaybeResult<T>& other) = default;
-  MaybeResult(MaybeResult<T>&& other) = default;
-  MaybeResult(const Err& e) : val(std::in_place_type<Err>, e) {}
-  MaybeResult(Err&& e) : val(std::in_place_type<Err>, std::move(e)) {}
-  template<typename U = T>
-  MaybeResult(U&& u) : val(std::in_place_type<T>, std::forward<U>(u)) {}
-  template<typename U = T>
-  MaybeResult(Result<U>&& u)
-    : val(u.getErr() ? std::variant<T, None, Err>{*u.getErr()}
-                     : std::variant<T, None, Err>{*u}) {}
-
-  // Whether we have an error or a value. Useful for assignment in loops and if
-  // conditions where errors should not get lost.
-  operator bool() const { return !std::holds_alternative<None>(val); }
-
-  Err* getErr() { return std::get_if<Err>(&val); }
-  T& operator*() { return *std::get_if<T>(&val); }
-  T* operator->() { return std::get_if<T>(&val); }
-
-  T* getPtr() { return std::get_if<T>(&val); }
-};
 
 // Parse a single WAT module.
 Result<> parseModule(Module& wasm, std::string_view in);
